working on prefix and typedId-s
